+++ projeto-tfg/src/main/kotlin/school/sptech/projetotfg/domain/cadastro/Contato.kt
@@ -11,15 +11,15 @@ import java.time.LocalDateTime
 
 @Entity
 class Contato(
-    @field:Id @field:GeneratedValue(strategy = GenerationType.IDENTITY)  var idContato:Int = 0,
+    @field:Id @field:GeneratedValue(strategy = GenerationType.IDENTITY)  var idContato:Long = 0,
     @field:PastOrPresent private var dataCriacao:LocalDateTime,
     @field:PastOrPresent private var dataUltimaAtualizacao:LocalDateTime,
     @field:Email @field:Size(max = 150) private var emailModificador:String
     ) {
-    fun getId():Int{
+    fun getId():Long{
         return idContato
     }
-    fun setId(novoId:Int){
+    fun setId(novoId:Long){
         idContato = novoId
     }
     fun setDataCriacao(novaDataCriacao:LocalDateTime){
