+++ projeto-tfg/src/main/kotlin/school/sptech/projetotfg/domain/gerenciamento/NivelAcesso.kt
@@ -5,13 +5,13 @@ import jakarta.validation.constraints.Max
 
 @Entity
 class NivelAcesso (
-    @field:Id @field:GeneratedValue(strategy = GenerationType.IDENTITY) private var idNivelAcesso:Int = 0,
+    @field:Id @field:GeneratedValue(strategy = GenerationType.IDENTITY) private var idNivelAcesso:Long = 0,
     @field:Max(60) private var apelido:String
 ){
-    fun getId():Int{
+    fun getId():Long{
         return idNivelAcesso
     }
-    fun setId(novoId:Int){
+    fun setId(novoId:Long){
         idNivelAcesso = novoId
     }
     fun getApelido():String{
